[tools] Radical refactoring on inputs for pca-out-of-core
It assumes that c3d features of one set (train, val, test) of a
dataset are in one HDF5-file. It also computes the mean and the matrix
X.T * X assuming that we can read all the features of a video in one
pass.
Note: this change break the compatibility with previous input.